pyproject.toml
[project]
name = "lakebase-foreachwriter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Christopher Grant", email = "chris.grant@databricks.com" }
]
requires-python = ">=3.11"
dependencies = [
    "databricks-connect>=16.1.0",
    "databricks-sdk==0.57.0",
    "psycopg==3.2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock",
    "ruff>=0.12.2",
    "ty>=0.0.1a14",
    "python-dotenv>=1.1.1",
]
